Further refactoring to consolidate Environment
